+++ HanBang_merging_1208/src/hanbang/controller/ReviewController.java
@@ -3,7 +3,6 @@ package hanbang.controller;
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
-import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
@@ -18,10 +17,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.ModelAndView;
 
 import hanbang.domain.Answer;
+import hanbang.domain.Member;
 import hanbang.domain.Review;
+import hanbang.service.MemberService;
 import hanbang.service.ReviewService;
 
 //@MultipartConfig(maxFileSize=1024*1024*1024, location="C:/Users/limsuhyun/eclipse-workspace/HanBang/WebContent/file")
@@ -30,6 +30,8 @@ public class ReviewController {
 
 	@Autowired
 	private ReviewService service;
+	@Autowired
+	private MemberService memberService;
 
 	// 후기 등록
 	@RequestMapping("/review/registReview.do")
@@ -39,23 +41,22 @@ public class ReviewController {
 
 	//
 	@RequestMapping(value = "/review/registReview.do", method = RequestMethod.POST)
-	public String uploadFile(@RequestParam("file") MultipartFile imgFile, Model model, Review review) {
+	public String uploadFile(@RequestParam("file") MultipartFile imgFile, Model model, Review review, HttpSession session) {
 
-		// String savePath ="C:/Users/limsuhyun/eclipse-workspace/HanBang_11.29_0.03/WebContent/uploadFile";
-		String savePath = "C:/Users/kosta/git/Final/HanBang/WebContent/uploadFile";
-		// String savePath = request.getRealPath("folderName"); // 파일이 저장될 프로젝트 안의 폴더경로
-
-		String originalFilename = imgFile.getOriginalFilename(); // fileName.jpg
-		System.out.println("originalFileName:" + originalFilename);
+		
+		String memberId = (String)session.getAttribute("memberId");
+		Member member = memberService.find(memberId);
+		
+		String savePath = "C:/Users/limsuhyun/git/HanBang/HanBang/WebContent/uploadFile";
+		String originalFilename = imgFile.getOriginalFilename() + ""; // fileName.jpg
 		String onlyFileName = originalFilename.substring(0, originalFilename.indexOf(".")); // fileName
 		String extension = originalFilename.substring(originalFilename.indexOf("."));
-		// .jpg
 
 		String rename = onlyFileName + "_" + getCurrentDayTime() + extension; // fileName_20150721-14-07-50.jpg
 		String fullPath = savePath + "\\" + rename;
 
-		review.setWriterId("ms");
-		review.setShareHouseId(1);
+		review.setWriterId(memberId);
+		review.setShareHouseId(member.getMemberTypeId());
 		review.setPhoto(rename);
 
 		service.register(review);
@@ -79,46 +80,6 @@ public class ReviewController {
 		return "redirect:detailReview.do?reviewId=" + reviewId;
 	}
 
-	// @RequestMapping(value = "review/registReview.do", method =
-	// RequestMethod.POST)
-	// public String uploadFile(@RequestParam("file") MultipartFile file, Model
-	// model, Review review) {
-	// if (!file.isEmpty()) {
-	// try {
-	// byte[] bytes = file.getBytes();
-	// File dir = new File("c:\\" + File.separator + "tempFiles");
-	//
-	// if (!dir.exists()) {
-	// dir.mkdirs();
-	// }
-	//
-	// File saveFile = new File(dir.getAbsolutePath() + File.separator + file);
-	// BufferedOutputStream out = new BufferedOutputStream(new
-	// FileOutputStream(saveFile));
-	// out.write(bytes);
-	// out.close();
-	//
-	// model.addAttribute("photo", "/images/" + file);
-	// } catch (IOException e) {
-	// e.printStackTrace();
-	// }
-	// }
-	// String originalFilename = file.getOriginalFilename(); // fileName.jpg
-	// String onlyFileName = originalFilename.substring(0,
-	// originalFilename.indexOf(".")); // fileName
-	// String extension = originalFilename.substring(originalFilename.indexOf("."));
-	// String rename = onlyFileName + "_" + getCurrentDayTime() + extension; //
-	//
-	// review.setWriterId("ms");
-	// review.setShareHouseId(1);
-	// review.setPhoto(rename);
-	//
-	// service.register(review);
-	// model.addAttribute(review);
-	//
-	// int reviewId = review.getReviewId();
-	// return "redirect:detailReview.do?reviewId=" + reviewId;
-	// }
 
 	// 시간정보
 	public String getCurrentDayTime() {
@@ -140,46 +101,72 @@ public class ReviewController {
 	// 후기 수정
 	@RequestMapping("/review/modifyReview.do")
 	public String modifyReview(int reviewId, HttpSession session, Model model) {
-		String memberId = (String) session.getAttribute("loginedUser");
+		String memberId = (String) session.getAttribute("memberId");
 		Review review = service.find(reviewId);
 		if (review.getWriterId().equals(memberId)) {
-			model.addAttribute(review);
-			return "modifyReview.jsp";
+			model.addAttribute("review", review);
+			return "/views/reviewModify.jsp";
 		} else {
-			return "reviewDetail.do";
+			return "redirect:reviewDetail.do?reviewId=" + reviewId;
 		}
 	}
 
 	//
 	@RequestMapping(value = "/review/modifyReview.do", method = RequestMethod.POST)
-	public String modifyReview(Review review, Model model, MultipartFile file) {
-		service.modify(review);
-		model.addAttribute(review);
-		return "detailReview.do";
+	public String modifyReview(int reviewId, Review review, Model model) {
+		try {
+			boolean check = service.modify(review);
+			if (check == false) {
+				model.addAttribute("review", review);
+				return "redirect:modifyReview.do?reviewId=" + reviewId;
+			} else {
+				model.addAttribute(review);
+				return "detailReview.do?reviewId=" + reviewId;
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+			return "redirect:modifyReview.do?reviewId=" + reviewId;
+		}
 	}
 
 	// 후기 삭제(reviewId)
 	@RequestMapping("/review/removeReviewByReviewId.do")
-	public String removeReviewByReviewId(int reviewId) {
-		boolean check = service.removeByReviewId(reviewId);
-		if (check == false) {
-			return "redirect:reviewDetail.do";
+	public String removeReviewByReviewId(int reviewId, HttpSession session) {
+		String memberId = (String) session.getAttribute("memberId");
+		Review review = service.find(reviewId);
+		if (review.getWriterId().equals(memberId)) {
+			boolean check = service.removeByReviewId(reviewId);
+			if (check == false) {
+				return "redirect:detailReview.do?reviewId=" + reviewId;
+			} else {
+				// shareHouseDetail.do에서 reviewList 뿌려줘야 함!!!!!!!!!
+				return "redirect:reviewList.do";
+			}
 		} else {
-			return "redirect:reviewList.do";
+			return "redirect:detailReview.do?reviewId=" + reviewId;
 		}
 	}
 
 	// 후기 삭제(memberId)
 	@RequestMapping("/review/removeReviewByMemberId.do")
-	public String removeReviewByMemberId(String memberId) {
-		boolean check = service.removeByMemberId(memberId);
-		if (check == false) {
-			return "redirect:reviewDetail.do";
+	public String removeReviewByMemberId(int reviewId, HttpSession session) {
+		String memberId = (String)session.getAttribute("memberId");
+		Review review = service.find(reviewId);
+		
+		if(review.getWriterId().equals(memberId)) {
+			boolean check = service.removeByMemberId(memberId);
+			if (check == false) {
+				return "redirect:reviewDetail.do";
+			} else {
+				// shareHouseDetail 에서 보여줘야하는 뷰!!!!!!!
+				return "redirect:reviewList.do";
+			}
 		} else {
-			return "redirect:reviewList.do";
+			return "redirect:/detailReview.do?reviewId=" + reviewId;
 		}
 	}
 
+	// ShareHouseController.
 	// 후기 삭제(shareHouseId)
 	@RequestMapping("/review/removeReviewByShareHouseId.do")
 	public String removeReviewShareHouseId(int shareHouseId) {
@@ -194,19 +181,66 @@ public class ReviewController {
 	// 후기 신고
 	@RequestMapping("/review/reportReview.do")
 	public String reportReview(HttpSession session, int reviewId) {
-//		String memberId = (String) session.getAttribute("loginedUser");
-		String memberId = "ms";
+		// String memberId = (String) session.getAttribute("loginedUser");
+		String memberId = "rr"; 
+
 		try {
 			boolean check = service.reportReview(memberId, reviewId);
-			if(check == false) {
+			System.out.println("*check : " + check);
+			if (check == false) {
+				// System.out.println(" WHY FALSE ? ");
 				return "redirect:detailReview.do?reviewId=" + reviewId;
 			} else {
-				return "redirect:detailReview.do?reviewId=" + reviewId;
+				// System.out.println("******Review REPORTED ! ");
+				// 후기 숫자 확인 후 자동 삭제
+				List<Integer> list = service.countReportReview(reviewId);
+				// System.out.println("@@@@reviewReportCount.size()" + list.size());
+				if (list.size() >= 2) {
+					boolean removeReports = service.removeReportedReviews(reviewId);
+					if (removeReports == false) {
+						// System.out.println(" REMOVE REPORTS FALSE");
+						return "redirect:registReview.do";
+					} else {
+						service.removeByReviewId(reviewId);
+						// System.out.println("**************Review REPORTED &&&&& DELETED BY REVIEWID
+						// !");
+						return "redirect:registReview.do";
+					}
+				} else {
+					// System.out.println("REVIEW REPORTED AND REDIRECTED ! *************8");
+					return "redirect:detailReview.do?reviewId=" + reviewId;
+				}
 			}
 		} catch (Exception e) {
-//			System.out.println(" ***** constraint violated !!");
-			return "redirect:detailReview.do?reviewId=" + reviewId;
+			System.out.println(" report review EXCEPTION !");
+			e.printStackTrace();
+			return "redrect:detailReview.do?reviewId=" + reviewId;
 		}
+		// try {
+		// boolean check = service.reportReview(memberId, reviewId);
+		// System.out.println("*check : " + check);
+		// if(check == false) {
+		// System.out.println(" WHY FALSE ? ");
+		// return "redirect:detailReview.do?reviewId=" + reviewId;
+		// } else {
+		// System.out.println("******Review REPORTED ! ");
+		// // 후기 숫자 확인 후 자동 삭제
+		// List<Integer> list = service.countReportReview(reviewId);
+		// System.out.println("@@@@reviewReportCount.size()" + list.size());
+		// if(list.size() >= 2) {
+		// service.removeByReviewId(reviewId);
+		// System.out.println("**************Review REPORTED &&&&& DELETED BY REVIEWID
+		// !");
+		// return "redirect:registReview.do";
+		// } else {
+		// System.out.println("REVIEW REPORTED AND REDIRECTED ! *************8");
+		// return "redirect:detailReview.do?reviewId=" + reviewId;
+		// }
+		// }
+		// } catch (Exception e) {
+		// System.out.println(" ***** constraint violated !!");
+		// return "redirect:detailReview.do?reviewId=" + reviewId;
+		// }
 	}
 
 }
